Finialized Advent of Code Day 13.
Part one was really easy... It took me a while to figure out the trick in part 2... It had to do with how that all the folds but one are on split by the middle most element. There is one fold that's on an even element. So that was some math to figure out the fold. There is a problem with the code though. It assumes that the top side or the left side are bigger respectively in the case of an even. If there is an uneven fold where the grid isn't split in half then it doesn't work. BUT this would be easily fixable. If there is an uneven fold, we just take the larger of the two halves and that should be the left or top side given this is transparent paper.

For the code itself I created a:
FoldableGrid class. I'm using a type vector to represent the grids. The reason is we are doing a bunch of index based lookups. Lists are very very slow for index based lookups O(N). Where as vector based lookups are eC (effectively constant). I also reused the point class container. I created a nice helper method to print off the elements in part 1 which paid dividends in part two.
a) A parse method that would take the input of points, then go over the points and create a grid from those coordinates.
b) Grid creation relied on knowing the max X and Y values to figure out the dimensions to create. I used a foldLeft to calculate the max values. I added all the points to a set and did a map to update all the points in the Array.
c) Created a helper function to parse the folds.
d) I have a step function that folds items one at a time. This could be called in a List.foldLeft.
e) Badly named but apt foldLeft, but this is in terms of folding the grid from left to right. This method goes over all the rowsthen only takes the Point.x items in the row. Then as we go from left to full right we check if the far left is visible OR take from the end - index + diff. This ensures that the values overlap.
f) foldUP does the same thing except on the Y axis. For both foldup and foldleft we look at the current value | flipped value.