Did the first part of Advent of Code Day 15.
I created a class ChitonPath that handles all the logic. I tried a really bad brute force method that kep ttrack of each route in a PriorityQueue of lists to slowly go through and find the shortest route... The problem was I was just having the different nodes and kept summing the routes in each list. This was VERY VERY slow. It took 25s on the TEST dataset and minutes too many minutes really on the full dataset.Most of the time was spent  on trying to make sure we weren't going backwards. That's when I sat down with a pen and paper and realized you could visit all nodes once. I even looked it up and found Djisktra / A* Algorithms. *Facepalm* I got the gist just from wikipedia. So this is my implementation from start to stop. I think with Djikstra you go to all nodes and not just start and stop?

Either way the main algorithm is in like 20 lines in the findShortestPath. I did do it with tail recursion which is nice. I added a lot of helper methods to clean up the code and make it more domain specific. I even used a visited map to figure out the cost to each node.