For Part 2 it was around extending the grids that we had. I created three helper methods. This required some refactoring because I now needd to have inputs for a ValueGrid datatype. This is what's used to create the matrix.
The first method is extending the rows. Which takes the original matrix and extending all the rows by width * extendSize.
The second method was extending the columns which takes a grid and extends all columns by height * extendedSize
I have a final method that calls both of these. It doesnt' matter which order that it happens given that the operation for extending a grid by rows and columns is associative. I wrote a nice little test to validate this.

Then I needed to do some refactoring. The parseINput became quite big so I reduced the size of that. I separarated out the creation of the matrix and the creation of the grid separately to make the code easier to read. I extend the grid by 1 as a default. This is a no-op. Adding in a second parameter that controls how large the matrix should eb on creation.

The other small class I created along with this was a rate limiter class. I was actually really curious how to implement a RateLimiter in scala.