I had the original test case done in an hour, but didn't account for the first part of the input algorithm turning flicking the outer pixels on and off. Once I realized that I changed up the algorithm. I got most of it done pretty quickly, but had an off by one error due to mixing up the border conditions :facepalm:
The ImageEnhancement class has all the main components. There are two types that I use within it. ImageAlgorithm which is just a boolean vector to see which portion of the algorithm will turn a light on or off. Then the ImageLocation is an Int Tuple. This is just x, y coordinates that was easier to model as a type. The one typeclass helper I created was a MinMaxDimensions holder. THis could have been a tuple of a tuple but this is what messaged up an off by one error. I made this for sanity.

Within the ImageENhancement Algorithm there is a method / lazy val that calculates the border dimensions. It does a foldLeft across all the current active pixels and finds the min and max values of each coordinate. This gives us the image border.
To handle the edge case whern the 000000000 binary algorithm flicks a pixel on I have a method that checks the current iteration and if it's odd and the first binary represenation has # then infinity is active. I have a helper method to check if a value is beyond the border. THis is really for when infinity is active.

The activePixels are held in a Set. This is just a way to handle the infinite growing size. It's also really quick to see if a neighbor is active or not.

Helper Method to get the points for all neighbors.

The applyEnhancementAlgorithm is the big workhorse. It grabs the current border... Then it has a modifier if we are on an odd iteration (e.g. when infinity is active) then the modifier is -2 / +2 else the modifier is +1 / -2... We iterate over all the border pixels and update the imagelocation for that border. This is done by taking all the neighbors of that pixel and checking whether or not they are on or off and building a binary representation from left to right top to bottom and looking up in the algorithm index.